Added a checkrecipe function
now a public function can be called from wherever the player places the finished plate to be checked against the generated recipe. the function returns a bool, true if the recipe is correct or false if not. So just call the function when the plate is submitted and then set the public bool "recipe_matched" from recipe gen script to award the player time and score if it matches